+++ apps/api/src/app.module.ts
@@ -1,26 +1,10 @@
 import { Module } from '@nestjs/common';
-import { APP_GUARD } from '@nestjs/core';
-import { AuthController } from './auth.controller';
-import { AuthRepository } from './auth.repository';
-import { SessionGuard } from './auth';
-import { AuthService } from './auth.service';
-import { DatabaseConnection } from './database.connection';
-import { DatabaseMigrator } from './database.migrator';
-import { StoriesController } from './stories.controller';
-import { StoriesRepository } from './stories.repository';
-import { StoriesService } from './stories.service';
+import { AuthModule } from './auth/auth.module';
+import { ConfigModule } from './config/config.module';
+import { DatabaseModule } from './database/database.module';
+import { StoriesModule } from './stories/stories.module';
 
 @Module({
-  controllers: [AuthController, StoriesController],
-  providers: [
-    AuthRepository,
-    AuthService,
-    DatabaseConnection,
-    DatabaseMigrator,
-    StoriesRepository,
-    StoriesService,
-    SessionGuard,
-    { provide: APP_GUARD, useExisting: SessionGuard },
-  ],
+  imports: [ConfigModule, DatabaseModule, AuthModule, StoriesModule],
 })
 export class AppModule {}